foodgram/tags/views.py
from rest_framework import viewsets
from tags.models import Tags
from tags.serializers import TagsSerializer


class TagsViewSet(viewsets.ReadOnlyModelViewSet):
    queryset = Tags.objects.all()
    serializer_class = TagsSerializer
