requirements.txt
Django==2.2.19
pytz==2022.2.1
sqlparse==0.4.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0